Add a spread support hook
This hook gives us an ubuntu/ubuntu user that can sudo without a password.
The system will boot up with DHCP network and SSH enabled out of the box.

We obviously don't want this (eventually) but it is a simple thing to
do from a hook. Later on this should be done in a step that occurs after
the snap is built (ideally just hack the user and enable stuff by default,
packages should be pre-installed).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>