static: add new core18.start-snapd.service
The new core18.start-snapd.service manages the snapd from the snapd
snap. It will run the seeding (if neeeded) and otherwise try to
start the right snapd from the snapd snap.

This is the initial version that will be able to run the seeding
and an installed snapd. It also contains basic fallback logic
where it will try to start the previous snapd if the current one
fails to run.