run-snapd-from-snap: start snapd after seeding
When snapd is seeding and it finised seeding the snapd snap it
will restart itself. The run-snapd-from-snap script can just
start snapd normally at this point. By just using the "current"
symlink we also ensure that each new snapd will restart into
the current version. The downside of this is that it makes
rollback handling of snapd harder.

This requires https://github.com/snapcore/snapd/pull/5330 to
work but it does no harm even if #5330 is not merged.