hooks: ensure /etc/login.defs PATHs contain /snap/bin
In ubuntu 18.04 /etc/environment is no longer used for the PATH
when using "su". It is unclear if this is a desired change but
until the root cause is found this PR adds /snap/bin to the PATH
in /etc/login.defs which is read by "su".

This will also unbreak some snapd integration tests that use
'su -l -c "cmd" test'.