hooks: add FIXME for /etc/login.defs changes
The "su.c" file from shadow is buggy and reads the pam environment first
and then overrides with the login.defs environment. This will be fixed
when https://github.com/shadow-maint/shadow/pull/119 is merged or when
we switch to "su" from util-linux. Until this is the case we need the
workaround in hooks/022-setup-path.chroot.