hooks: add test that ensures "synced" dirs are no longer used
Also run hooks after static content.

The "synced" filemode for the writable-paths had resulted in subtle
bugs in the past. One issue is that it is not possible to delete
a file in "synced" mode, because writable-paths will always
re-create the file from the core18 image. There are more subtle
issues so we should not rely on this.