static: make run-snapd-from-snap wait for seeding
We need to wait in core18.run-snapd.service for snapd to be fully
seeded. This is required because console-conf runs after
core18.run-snapd.service but we cannot make it run after
snapd.seeded.service because at the time when the dependencies
for core18.run-snapd.service are calculated by systemd there is
no snapd.seeded yet.

It also redirects the output of the seeding to the console so
that its easier to see what is going on during seeding.