run-snapd-from-core: fix race when restarting snapd.socket
In the early core18 bootstrap there is no systemd owned
snapd.socket. The snapd daemon starts and creates this
socket itself. However when snapd restarts itself golang
will remove this socket (we have no control over this).

This may mean that the snapd.seeded job has the wrong
socket open (the socket that was owned by the bootstraping
snapd) and because there is no snapd anymore behind this
socket the snapd.seeded service fails.

To fix this the snapd.seeded service is also restarted
after the snapd.socket gets restarts. This ensures that
the snapd.seeded service will watch the right socket.