static: run snapd for the first time with systemd-run
So far we ran the snapd for the first time wihtout systemd. This
has the undesired side-effect that snapd creates the snapd.socket
file and when it exists this file gets removed by the golang runtime.

This causes all sorts of problems because the snapd.seeded service
tries to talk to a socket file that got removed. By using systemd
we get a socket file that is managed by systemd and not removed
which fixes the issues that snapd.seeded explodes.