static,hooks: make /etc/cloud a "synced" dir for now
We write an empty /writable/system-data/etc/cloud directory into
the UC18 image. This means that the code that processes the
"transition" dirs assumes the content was already copied and skips
this dir. Using "synced" unbreaks things.

Once we have removed the creation of this dir during image bulding
we can revert this patch again.