hooks: create snapd directory skeleton
Even when running from snapd snap we sometimes assume that certain
directories exist in the base snap where we implant ourselves. We need
the void directory for sure, because the tests showed it was missing and
it is infeasible for snap-confine to always create it. I suspect
something similar will happen to the GL directories but testing in that
area is partial, at best. The directory structure is modelled after
snapd.mk from the snapd tree's packaging/ directory. The list will be
duplicated now but it's better than not having those directories in the
first place.

An existing hook that was creating "classic directory" was removed, it
was creating the hostfs directory and I think it warrants that we make
all of the directories at once in one hook.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>