bootstrap from locally build snapd snaps too
An unasserted snapd snap will be located in
```
/var/lib/snapd/seed/systems/$label/snaps
```
which the current run-snapd-from-snap bootstrap code will not
find. This commit fixes this to make this work short-term.

However we will need to write the right snapd snap into the
grubenv/modeenv so that when there are multiple recovery systems
there is no confusion.