static: switch /etc/cloud from synced to persistent
This commit move /etc/cloud from a synced dir to a persistent
dir. Synced dirs have strange semantics, e.g. the user cannot
remove a file that is also present in the core20 snap in the
dir without it coming back on the next reboot. Hence we want
to avoid any synced dirs.

This requires the following prereqs:
https://github.com/snapcore/core20/pull/46
https://github.com/snapcore/snapd/pull/8612

first.