hooks: fix broken symlink /etc/sysctl.conf.d/99-sysctl.conf
Core currently has a symlink /etc/sysctl.conf.d/99-sysctl.conf that
points to /etc/sysctl.conf. This file is removed by the core build
so removing this symlink is also needed.

Note that the sysctl.conf file is not writable on core and it
only contains commented out entries.