.travis.yml: use snapcraft w/ lxd to build the snap
We can't use sudo snapcraft with snapcraft 4.0, and we should ideally be
building with lxd, as that provides us with the right build-base for building
the snap, regardless of what series travis currently supports.

Also we need snapcraft 4.1.2 to take advantage of snapcraft intercepting the 
mknod calls, otherwise the lxd container fails when trying to run the chroot
things.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>