Add a wrapper for invoking the snap binary
On Core systems with the snapd snap, /usr/bin/snap is a symlink pointing to
/snap/snapd/current/usr/bin/snap. However, while the snapd snap is being
refreshed, the current symlink becomes invalid for a short while. Should a
reboot occur while this happens, the existing snap services will be unable to
start as /usr/bin/snap points to an invalid location.

To address this, we move the snap binary to /usr/lib/snapd/snap, as
/usr/lib/snapd gets exported as the usr-lib-snapd.mount. Then the core18/20 snap
has /usr/bin/snap which is a wrapper script that picks
/snap/snap/current/usr/lib/snap or /usr/lib/snapd/snap.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>