hooks: enable systemd-time-wait-sync.service
The snapd team recently got bugreport about refresh issues that
look related to an incorrect time for devices without a battery
clock like the Raspberry Pi.

Upon inspection it looks like we do not enable the
systemd-time-wait-sync.service on core20 and this means that
we cannot ensure that snapd run only *after* the time is synced.
It seems right now we hit the race that timesyncd starts before
snapd but if it's slow at syncronizing snapd will start with the
wrong time.

Once we have systemd-time-wait-sync.service we can add:
```
After=time-sync.target
```
and that should eliminate this race.