hooks: mkdir /var/cups to avoid writable mimic creation for snaps using cups
Snaps using the cups interface with the cups snap as of snapd 2.55 will create
a bind mount of /run/cups -> /var/cups, which since /var/cups does not exist
will trigger the writable mimic code to create a writable mimic on /var. This
results in a fair amount of extra bind mounts for every directory in /var, so
creating this directory empty in the base snap ensures that no writable mimic
need be created.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>